Added note regarding presto-cli vs presto
The EMR Presto documentation references the Presto documentation, but nowhere mentions that to run Presto from the command line you need to use "presto-cli" vs. the normal "presto". I spent a good amount of time beating my head against this one and eventually opened a support request to get the answer. Someone familiar with Presto would probably know this or be able to discover it, but I'm coming from a Cloudera (Impala) background. This item would have saved me an hour.