Update emr-jupyterhub-ldap-users.md
Added a section with a list of attributes that must exist for the JupyterHub container PAM to find the user on Active Directory. They are called Unix attributes, and they are not created by default when a new user is created. I tested it with AWS managed AD and EMR emr-5.19.0. Without those attributes the JupyterHub is able to authenticate the user, but then generate an internal error (500) because the Linux PAM is not able to get the user information from LDAP, such as the user's home directory where JupyterHub stores the notebook data.